@@ -4,50 +4,51 @@
   "private": true,
   "proxy": "http://localhost:8080",
   "dependencies": {
-    "@tailwindcss/aspect-ratio": "^0.2.0",
-    "@tailwindcss/typography": "^0.3.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
+    "@tailwindcss/aspect-ratio": "^0.4.0",
+    "@tailwindcss/typography": "^0.5.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.2.1",
     "@types/classnames": "^2.2.10",
-    "@types/jest": "^24.0.0",
-    "@types/js-cookie": "^2.2.6",
+    "@types/jest": "^28.1.3",
+    "@types/js-cookie": "^3.0.2",
     "@types/lodash": "^4.14.165",
-    "@types/node": "^12.0.0",
+    "@types/node": "^18.0.0",
     "@types/query-string": "^6.3.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
+    "@types/react": "^17.0.2",
+    "@types/react-dom": "^17.0.2",
     "@types/react-router-dom": "^5.1.5",
     "@types/react-table": "^7.0.25",
     "@types/react-transition-group": "^4.4.0",
-    "@types/yup": "^0.29.9",
-    "@typescript-eslint/eslint-plugin": "^4.8.2",
-    "@typescript-eslint/parser": "^4.8.2",
-    "@xstate/immer": "^0.1.0",
-    "@xstate/react": "^1.0.1",
+    "@types/yup": "^0.29.14",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
+    "@xstate/immer": "^0.3.1",
+    "@xstate/react": "^3.0.0",
     "autoprefixer": "^10.0.2",
     "classnames": "^2.2.6",
     "formik": "^2.2.1",
-    "framer-motion": "^2.9.3",
-    "graphql": "^15.3.0",
-    "graphql-request": "^3.4.0",
+    "framer-motion": "^6.3.13",
+    "graphql": "^16.5.0",
+    "graphql-request": "^4.3.0",
     "immer": "^9.0.6",
-    "js-cookie": "^2.2.1",
+    "js-cookie": "^3.0.1",
     "lodash": "^4.17.20",
-    "postcss": "^8.2.10",
-    "query-string": "^6.13.8",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.1",
-    "react-table": "^7.6.2",
-    "react-toastify": "^6.2.0",
-    "tailwindcss": "~2.0.0",
-    "typescript": "^4.1",
-    "use-immer": "^0.6.0",
-    "xlsx": "^0.18.2",
-    "xstate": "^4.23.1",
-    "yup": "^0.29.3"
+    "postcss": "^8.4.14",
+    "query-string": "^7.1.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "6",
+    "react-router-navigation-prompt": "^1.9.6",
+    "react-scripts": "^5.0.1",
+    "react-table": "^7.8.0",
+    "react-toastify": "^9.0.5",
+    "tailwindcss": "~3.1.4",
+    "typescript": "^4.7.4",
+    "use-immer": "^0.7.0",
+    "xlsx": "^0.18.5",
+    "xstate": "^4.32.1",
+    "yup": "^0.32.11"
   },
   "scripts": {
     "start": "npm-run-all -p start:css start:js",
