@@ -1,6 +1,9 @@
-import React from "react";
+import React, { ReactNode } from "react";
 
-const ButtonBar: React.FC = ({ children }) => {
+type ButtonBarProps = {
+  children: ReactNode;
+};
+const ButtonBar = ({ children }: ButtonBarProps) => {
   return (
     <div className="border border-t-2 border-gray-200 w-full py-4 px-4 sm:px-6 lg:px-8 bg-gray-100 flex-shrink-0">
       <div className="max-w-screen-xl mx-auto">
