@@ -26,8 +26,9 @@ import FormikInput from "../../forms/Input";
 import PotProcessingLabwarePlan from "./PotProcessingLabwarePlan";
 import ProcessingSuccess from "../ProcessingSuccess";
 import columns from "../../dataTable/labwareColumns";
-import { Prompt } from "react-router-dom";
 import { useConfirmLeave } from "../../../lib/hooks";
+import NavigationPrompt, { ChildData } from "react-router-navigation-prompt";
+import { ConfirmationNavigationModal } from "../../modal/ConfirmationNavigationModal";
 
 /**
  * Used as Formik's values
@@ -249,7 +250,7 @@ export default function PotProcessing({ processingInfo }: PotProcessingParams) {
   /**
    * Builds a yup validator for the labware plan form
    */
-  function buildValidationSchema(): Yup.ObjectSchema {
+  function buildValidationSchema() {
     return Yup.object().shape({
       workNumber: Yup.string().required(),
       sourceBarcode: Yup.string().required(),
@@ -414,10 +415,16 @@ export default function PotProcessing({ processingInfo }: PotProcessingParams) {
           </Formik>
         )}
       </motion.div>
-      <Prompt
-        when={shouldConfirm}
-        message={"You have unsaved changes. Are you sure you want to leave?"}
-      />
+
+      {/**This is the replacement of Prompt which is removed in react-router version 6**/}
+      <NavigationPrompt when={shouldConfirm}>
+        {({ onConfirm, onCancel }: ChildData) => (
+          <ConfirmationNavigationModal
+            onConfirm={onConfirm}
+            onCancel={onCancel}
+          />
+        )}
+      </NavigationPrompt>
     </>
   );
 }
