@@ -241,10 +241,8 @@ function Extraction() {
           </Link>
           <div className={""}>
             <Link
-              to={{
-                pathname: "/lab/extraction_result",
-                state: { labware: extraction?.extract?.labware },
-              }}
+              to="/lab/extraction_result"
+              state={{ labware: extraction?.extract?.labware }}
             >
               <BlueButton action="primary">
                 Go to Extraction Result &gt;
